+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 
 * Always parse imports with queries as plain CSS imports.
 
+* Support `&` followed by a non-identifier.
+
 ## 1.0.0-alpha.8
 
 * Add the `content-exists()` function.